Delete README.md
# British Airways Customer Feedback Analysis
## Background
British Airways (BA) is the flag carrier airline of the United Kingdom (UK). As a data scientist at BA, it is important to understand customer feelings, needs, and feedback to improve the business and provide top-class customer service.

This project involves scraping and collecting customer feedback data from the website Skytrax.com and analyzing the data to uncover insights about the airline.

## Problem Statement
Understanding customer feelings, needs, and feedback is crucial for improving business and providing top-class customer service. British Airways (BA) is interested in analyzing customer review data to uncover insights about the airline.

## Objectives
The main objectives of this project are:

1. Scrape customer review data from Skytrax.com and clean and prepare the data for analysis.
2. Perform data analysis on the cleaned data using techniques such as topic modeling, sentiment analysis, and wordclouds to uncover insights about the content of the reviews.
3. Create a PowerPoint slide with visualizations and metrics to summarize the key findings of the analysis, along with clear and concise explanations.

### FINAL PRESENTATION
![Task 1 - Example Answer updated-2](https://user-images.githubusercontent.com/115629197/207058693-6a40e3f7-5835-4578-bf63-e4d419215dbf.png)

## INSIGHTS
### WORD CLOUD
![word cloud](https://user-images.githubusercontent.com/115629197/207290538-aa369a06-c0f4-4c50-9c38-784161c342a3.png)

### HISTOGRAM
![hist](https://user-images.githubusercontent.com/115629197/207290771-f50e6c5f-78ec-47a9-b724-b1adc3807d90.png)

### PIE CHART
![pie](https://user-images.githubusercontent.com/115629197/207290790-bada8f1b-8822-4f70-8e9b-3cf1b781d1da.png)
## Data
The data for this project will be collected from Skytrax.com and will consist of customer reviews about the airline. The data will be in the form of text and will require cleaning and preparation before it can be analyzed.

## Resources
The following resources will be used in this project:

1. Skytrax.com for customer review data
2. Python and libraries such as BeautifulSoup and Pandas for data scraping and preparation
3. Techniques such as topic modeling, sentiment analysis, and wordclouds for data analysis
4. PowerPoint for presenting the results and recommendations
## Methodology
The following steps will be taken to complete this project:

1. Data scraping and preparation: Scrape customer review data from Skytrax.com and clean and prepare the data for analysis using Python and libraries such as BeautifulSoup and Pandas.
2. Data analysis: Perform data analysis on the cleaned data using techniques such as topic modeling, sentiment analysis, and wordclouds to uncover insights about the content of the reviews.
3. Results and recommendations: Create a PowerPoint slide with visualizations and metrics to summarize the key findings of the analysis, along with clear and concise explanations